+++ core/src/test/java/io/grpc/internal/ManagedChannelImplIdlenessTest.java
@@ -62,7 +62,7 @@ import io.grpc.NameResolver.ResolutionResult;
 import io.grpc.Status;
 import io.grpc.StringMarshaller;
 import io.grpc.internal.FakeClock.ScheduledTask;
-import io.grpc.internal.ManagedChannelImplBuilder.ClientTransportFactoryBuilder;
+import io.grpc.internal.ManagedChannelImplBuilder.UnsupportedClientTransportFactoryBuilder;
 import io.grpc.internal.TestUtils.MockClientTransportInfo;
 import java.net.SocketAddress;
 import java.net.URI;
@@ -161,12 +161,7 @@ public class ManagedChannelImplIdlenessTest {
         .thenReturn(timer.getScheduledExecutorService());
 
     ManagedChannelImplBuilder builder = new ManagedChannelImplBuilder("fake://target",
-        new ClientTransportFactoryBuilder() {
-          @Override public ClientTransportFactory buildClientTransportFactory() {
-            throw new UnsupportedOperationException();
-          }
-        },
-        null);
+        new UnsupportedClientTransportFactoryBuilder(), null);
 
     builder
         .nameResolverFactory(mockNameResolverFactory)
